git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/branches/b_djl_NoteText@565 41062d70-b26d-460a-9302-f438a8930db8

+++ src/DPoint.h
@@ -65,8 +65,8 @@ public:
 	CDPoint()
 	{
 		// Initialize with not-a-number
-		x = std::numeric_limits<long double>::quiet_NaN();
-		y = std::numeric_limits<long double>::quiet_NaN();
+		x = std::numeric_limits<double>::quiet_NaN();
+		y = std::numeric_limits<double>::quiet_NaN();
 	}
 
 	CDPoint(double ix, double iy)
